% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjMat.R
\name{adjMat}
\alias{adjMat}
\title{Spatial adjacency matrix}
\usage{
adjMat(m, n, boundary = FALSE)
}
\arguments{
\item{m}{height of the grid}

\item{n}{width of the grid}

\item{boundary}{surround grid by an external boundary, defaults to FALSE}
}
\value{
a list containing a sparse adjacency matrix for a square lattice, and the vector of the points not on the edge
}
\description{
Takes as input a matrix gridMat with the following format: 0 means a site is included in the study area, and nonzero values mean the site is not included.
Returns an adjacency matrix for the study sites, where sites at a Manhattan distance of exactly kNN are considered neighbors.
}
\examples{
mat=adjMat(10,10)
}
\keyword{adjacency,}
\keyword{grid}
\keyword{neighborhood,}
\keyword{spatial,}

