% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indsFunction.R
\name{indsFunction}
\alias{indsFunction}
\title{Get indices of relevant component of theta matrix}
\usage{
indsFunction(theta, X, z, A)
}
\arguments{
\item{theta}{matrix of coefficients}

\item{X}{design matrix}

\item{z}{matrix of multinomial responses}

\item{A}{list of adjacency matrices}
}
\value{
list of index vectors
}
\description{
Gets indices of beta, eta, diagonal of eta, and off diagonal elements of eta
}
\examples{
#
}
\keyword{centered}

