% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMatrixCentered.R
\name{JMatrixCentered}
\alias{JMatrixCentered}
\title{J matrix for centered variance estimate}
\usage{
JMatrixCentered(theta, X, z, A, constraint, thetaInds)
}
\arguments{
\item{theta}{coefficients}

\item{X}{design matrix}

\item{z}{response matrix}

\item{A}{list of adjacency matrices}

\item{constraint}{"symmetric" or "diagonal"}

\item{thetaInds}{list of relevant indices in theta: betaInds, etaDiag, ...}
}
\value{
A sparse adjacency matrix
}
\description{
Computes J matrix for the centered variance estimate
}
\examples{
#JMatrix=JMatrixCentered(theta,X,z,A)
}
\keyword{adjacency,}
\keyword{grid}
\keyword{neighborhood,}
\keyword{spatial,}

