% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMatrixUncentered.R
\name{JMatrixUncentered}
\alias{JMatrixUncentered}
\title{Uncentered J matrix}
\usage{
JMatrixUncentered(theta, X, z, A, constraint, thetaInds, innerIndices)
}
\arguments{
\item{theta}{a matrix of coefficients}

\item{X}{a design matrix}

\item{z}{a matrix of responses}

\item{A}{a list of adjacency matrices}

\item{constraint}{"symmetric" or "diagonal"}

\item{thetaInds}{list of useful indices}

\item{innerIndices}{the 1-dimensional indices of points within the grid}
}
\value{
The J matrix for the sandwich variance estimator.
}
\description{
Used for a variance estimate accounting for correlation between sites or for sandwich estimate for independent sites.
}
\examples{
#JMatrix=JMatrixUnCentered(theta,X,z,A,innerIndices)
}
\keyword{autologistic,}
\keyword{uncentered,}
\keyword{variance}

