% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikelihood.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Compute log likelihood}
\usage{
logLikelihood(theta, X, z, A, constraint, thetaInds, innerIndices)
}
\arguments{
\item{theta}{a matrix of coefficinets}

\item{X}{design matrix}

\item{z}{response matrix}

\item{A}{list of adjacency matrices/list containing adjacency matrix}

\item{constraint}{constraint on eta}

\item{thetaInds}{relevant indices}

\item{innerIndices}{internal grid indices}
}
\value{
A sparse adjacency matrix
}
\description{
Compute log likelihood for uncentered or independent case models
}
\examples{
#logLikelihood(theta,X,z,A,constraint,thetaInds)
}
\keyword{adjacency,}
\keyword{grid}
\keyword{neighborhood,}
\keyword{spatial,}

