% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPLE_summary.R
\name{MPLE_summary}
\alias{MPLE_summary}
\title{Summarize MPLE fits}
\usage{
MPLE_summary(fit)
}
\arguments{
\item{fit}{a fitted MPLE object}
}
\value{
tables based on the model fit
}
\description{
Prints out summary tables of fitted model objects from MPLE. Also
returns knitr::kable() summary tables.
}
