% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialBCD.R
\name{multinomialBCD}
\alias{multinomialBCD}
\title{Block coordinate descent}
\usage{
multinomialBCD(X, z, groups, penaltyFactor, nLambda, H = NULL, theta = NULL)
}
\arguments{
\item{X}{design matrix}

\item{z}{response matrix}

\item{groups}{a grouping of coefficients}

\item{penaltyFactor}{a vector specifying the relative penalty level for each coefficient group}

\item{nLambda}{number of penalty parameters on the lasso path}

\item{H}{optional Hessian at loglikelihood maximum, used for the quadratic approximation in the centered model}

\item{theta}{optional coefficients at loglikelihood maximum, used for the quadratic approximation in the centered model}
}
\value{
list containing the lambdas on the coordinate descent path and the coefficients for each lambda
}
\description{
Computes, with group lasso regularization penalty, a lasso path for a model with user specified loglikelihood gradient and hessian.
}
\examples{
#
}
\keyword{block,}
\keyword{coordinate,}
\keyword{descent}

