% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multProb.R
\name{multProb}
\alias{multProb}
\title{Multinomial probabilities}
\usage{
multProb(X, beta)
}
\arguments{
\item{X}{nxp design matrix}

\item{beta}{a pxk coefficient matrix}
}
\value{
an nxk matrix of probabilities
}
\description{
Takes a design matrix and coefficients and computes the fitted probabilities for these coefficients.
}
\examples{
#probs=multProb(X,beta)
}
\keyword{fitted,}
\keyword{multinomial,}
\keyword{probabilities}

