% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Visualize response values on a grid}
\usage{
plotGrid(z, m, n)
}
\arguments{
\item{z}{nxk matrix or length n vector of responses, where z contains the responses of the grid in column-major order}

\item{m}{the number of rows in the grid}

\item{n}{the number of columns in the grid}
}
\value{
void
}
\description{
Takes in an nxk matrix or a length n vector containing the response type of each observation
and creates an image of the response for a specified grid size
}
\examples{
#plotGrid(z,m,n)
}

