% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwichVariance.R
\name{sandwichVariance}
\alias{sandwichVariance}
\title{Compute sandwich variance estimate}
\usage{
sandwichVariance(theta, X, z, A, centered, I = NULL, constraint, thetaInds,
  innerIndices)
}
\arguments{
\item{theta}{a coefficient matrix}

\item{X}{a design matrix}

\item{z}{a matrix of multinomial responses}

\item{A}{a list of adjacency matrices}

\item{centered}{use centered or uncentered model}

\item{I}{Fisher information matrix}

\item{constraint}{constraint on eta}

\item{thetaInds}{relevant indices}

\item{innerIndices}{internal grid indices}
}
\value{
a covariance matrix
}
\description{
Computes the sandwich variance estimate for a given neighborhood structure and coefficient estimate
}
\examples{
#
}
\keyword{adjacency,}
\keyword{grid}
\keyword{neighborhood,}
\keyword{spatial,}

