% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalX.R
\name{totalX}
\alias{totalX}
\title{Add columns to X matrix}
\usage{
totalX(theta, X, z, A)
}
\arguments{
\item{theta}{matrix of coefficients}

\item{X}{design matrix}

\item{z}{matrix of multinomial responses}

\item{A}{list of adjacency matrices}
}
\value{
an expanded design matrix
}
\description{
Adds the A%*%(z-mu) columns to the design matrix. Used for fitting the centered model
}
\examples{
#
}
\keyword{centered}

