% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialGradient.R
\name{uncenteredLogLikGrad}
\alias{uncenteredLogLikGrad}
\title{Uncentered gradient}
\usage{
uncenteredLogLikGrad(theta, X, z, A, constraint, thetaInds, innerIndices)
}
\arguments{
\item{theta}{matrix of coefficients}

\item{X}{an nxp design matrix}

\item{z}{an nxk response matrix}

\item{A}{list of adjacency matrices/list containing adjacency matrix}

\item{constraint}{constraint on eta}

\item{thetaInds}{relevant indices}

\item{innerIndices}{the (1-dim.) location/index for internal points in the grid}
}
\value{
a gradient vector
}
\description{
Takes coefficient values and a design matrix and computes the gradient of the loglikelihood
}
\examples{
#gradient=uncenteredLogLikGrad(theta,X,z,A,constraint,thetaInds)
}
\keyword{gradient}
\keyword{multinomial,}

