% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialHessian.R
\name{uncenteredLogLikHess}
\alias{uncenteredLogLikHess}
\title{Uncentered hessian}
\usage{
uncenteredLogLikHess(theta, X, z, A, constraint, thetaInds)
}
\arguments{
\item{theta}{matrix of coefficients}

\item{X}{an nxp design matrix}

\item{z}{response matrix}

\item{A}{a list of adjacency matrices/list containing an adjacency matrix}

\item{constraint}{constraint on eta}

\item{thetaInds}{relevant indices}
}
\value{
a hessian matrix
}
\description{
Takes coefficient values and a design matrix and computes the hessian matrix of the loglikelihood
}
\examples{
#hess=uncenteredLogLikHess(theta,X,z,A)
}
\keyword{multinomial,}
\keyword{regression}

