% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unStandardize.R
\name{unStandardize}
\alias{unStandardize}
\title{Unstandardize coefficients}
\usage{
unStandardize(thetaList, means, sds, nRow)
}
\arguments{
\item{thetaList}{a list of coefficient matrices}

\item{means}{means of the non-intercept columns on the original scale}

\item{sds}{standard deviations of the coefficients on the original scale}

\item{nRow}{the number of coefficient rows that are affected by standardization and unstandardization}
}
\value{
a list of coefficient matrices on the original scale
}
\description{
Converts coefficients on a standardized scale to the original scale
}
\examples{
#
}
\keyword{standardize}

