% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawSamples.R
\name{drawSamples}
\alias{drawSamples}
\title{Simulate data from auto- models}
\usage{
drawSamples(beta, gamma, X, A, burnIn = 300, nSamples)
}
\arguments{
\item{beta}{coefficient vector (for the autologistic model) or matrix (for the automultinomial model)}

\item{gamma}{the value of the autocorrelation parameter}

\item{X}{the design matrix}

\item{A}{the (square symmetric) adjacency matrix encoding the neighborhood structure}

\item{burnIn}{the number of burnin samples to be used. Defaults to 300}

\item{nSamples}{the number of samples to draw}
}
\value{
simulated samples
}
\description{
Generates data from the autologistic and automultinomial
models via Gibbs sampling. See the vignette for an example of use.
}
