% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multProb.R
\name{MPLE}
\alias{MPLE}
\title{Maximum pseudolikelihood estimation}
\usage{
MPLE(X, y, A, ciLevel = 0.95, method = "asymptotic", burnIn = 300,
  nBoot = 500)
}
\arguments{
\item{X}{the n-by-p design matrix}

\item{y}{the response vector (required to be a factor)}

\item{A}{the square symmetric adjacency matrix A encoding the neighborhood structure}

\item{ciLevel}{the confidence level to be used for inference. Defaults to 95 percent intervals.}

\item{method}{"boot" for parametric bootstrap and "asymptotic" for asymptotic confidence intervals.}

\item{burnIn}{the number of burnin samples to use for the Gibbs sampler when method="boot"}

\item{nBoot}{the number of bootstrap samples to use when method="boot"}
}
\value{
a fitted auto- model MPLE object
}
\description{
Fits an autologistic model or automultinomial model. Takes as arguments
a design matrix X, a response vector y (in factor form), and a square
symmetric adjacency matrix encoding the neighborhood structure.
When the number of levels of the response y is >2, the function fits the
multicategory generalization of the autologistic model. For a full description
of the models and a user guide, please see the vignette.
}
